all: Reorganize logic to better suit event based window handling
After exploring the original browsercontrol.dll code as well as how
win32 windows are created and manipulated there were a few things which
needed to be changed about how we were exposing the browserwindow to the
JNI interface.

This is the current (yet to be validated) plan:

JNI interface will be responsible for converting JNI types to std types
BrowserContext will be responsible for calling event creators
BrowserWIndow will be responsible for creating windows, setting up event
handlers, and manipulating windows.

Signed-off-by: Open592 Developer <uss@open592.com>