windows: Move browser window to own thread
We need to return after initializing the browser window. In order to do
this we need to offload the browser window to it's own thread and signal
when it has been successfully created and is beginning to receive
messages.

Once this is successful we will return back to the caller

Signed-off-by: Open592 Developer <uss@open592.com>