all: Add Browser BrowserContext and BrowserData and Windows impl
Refactored the way we were handling platform specific logic. Previously
our BrowserData class was used by all platforms and didn't allow for
platform specific data to be added without the addition of a seperate
class.

Now we provide a Base implementation of BrowserContext and BrowserData
which all supported platforms can extend to support their needs.
BrowserContext will wrap the data and provide access to it.

We also introduce the concept of a global "Browser" singleton which
holds the reference to the BrowserContext.

Signed-off-by: Open592 Developer <uss@open592.com>