windows: Remove WindowsBrowserContext::The()
No need to expose a "singleton" reference to the browser context.
Especially for Windows we can just pass the BrowserData to the
WebView2BrowserWindow. It's not the cleanest since we need to use the
weird Win32 APIs - but I'd like to use a similar approach for all
platforms, and for Linux I think I can figure out a cleaner solution
without having the Singleton The() method.

Signed-off-by: Open592 Developer <uss@open592.com>