linux: Decrease browser functionality
We don't want the browser to be showing context menus or creating popup
windows.

In the future when the browser requests to open a popup window (for
example, from a target="_blank" <a>) we will intercept that request and
forward it to the desktop's default browser. This was the behavor of the
browsercontrol within the AppletViewer.

Signed-off-by: Open592 Developer <uss@open592.com>