examples: Move invokation of setBounds()
Upon loading the example application the inner fakeApplet was not taking
up the total size of the frame. Upon resizing the window this fixed
itself.

I think there is some weird behavior on Linux with window bounds which I
will need to fix within the applet viewer. Once I do that I might have a
better way to fix this bug.

Signed-off-by: Open592 Developer <uss@open592.com>